crawler/requirements.txt
Scrapy~=2.11.0
schedule~=1.2.1
itemadapter~=0.8.0
requests~=2.31.0
typer~=0.9.0
kafka-python~=2.0.2
cassandra-driver==3.28.0
python-dotenv~=1.0.0
scrapyd==1.4.3
python-scrapyd-api==2.1.2